Layout Service/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy>=2.0
python-dotenv>=1.0
alembic>=1.13
psycopg2-binary>=2.9
